Add PAG Config page and optimize parameter I/O with TinyXMLexporter (#2903)

+++ exporter/src/utils/PsdTextAttribute.cpp
@@ -17,6 +17,7 @@
 /////////////////////////////////////////////////////////////////////////////////////////////////
 
 #include "PsdTextAttribute.h"
+#include <algorithm>
 #include <array>
 #include <charconv>
 #include <iostream>
@@ -35,8 +36,8 @@ inline int HexCharToInt(const char c) {
   return result;
 }
 
-int PsdTextAttribute::stringFormatTransform(std::vector<char>& dst, const uint8_t* src,
-                                            const int len) {
+size_t PsdTextAttribute::stringFormatTransform(std::vector<char>& dst, const uint8_t* src,
+                                               const int len) {
   dst.push_back('/');
   dst.push_back('0');
   dst.push_back(' ');
